+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Feature:
 Bug fix:
 - [#1235](https://github.com/elysiajs/elysia/issues/1235) errors not handled correctly in resolve hook on dynamic mode
 - [#1234](https://github.com/elysiajs/elysia/issues/1234) optional path parameters can't follow required ones
+- [#1232](https://github.com/elysiajs/elysia/issues/1232) t.Files fails with array of files
 
 Change:
 - When yield is not sse, content-type is set to either `text/plain` or `application/json` based on the response type